+++ gst/gl/gstglfilterblur.c
@@ -239,11 +239,11 @@ gst_gl_filterblur_draw_texture (GstGLFilterBlur * filterblur, GLuint tex)
 
   glTexCoord2f (0.0, 0.0);
   glVertex2f (-1.0, -1.0);
-  glTexCoord2f (filter->width, 0.0);
+  glTexCoord2f ((gfloat)filter->width, 0.0);
   glVertex2f (1.0, -1.0);
-  glTexCoord2f (filter->width, filter->height);
+  glTexCoord2f ((gfloat)filter->width, (gfloat)filter->height);
   glVertex2f (1.0, 1.0);
-  glTexCoord2f (0.0, filter->height);
+  glTexCoord2f (0.0, (gfloat)filter->height);
   glVertex2f (-1.0, 1.0);
 
   glEnd ();
@@ -290,9 +290,9 @@ gst_gl_filterblur_hcallback (gint width, gint height, guint texture, gpointer st
   /* hard coded kernel, it could be easily generated at runtime with a
    * property to change standard deviation */
   gfloat gauss_kernel[9] = {
-    0.026995, 0.064759, 0.120985,
-    0.176033, 0.199471, 0.176033,
-    0.120985, 0.064759, 0.026995 };
+    0.026995f, 0.064759f, 0.120985f,
+    0.176033f, 0.199471f, 0.176033f,
+    0.120985f, 0.064759f, 0.026995f };
 
   glMatrixMode (GL_PROJECTION);
   glLoadIdentity ();
@@ -307,8 +307,8 @@ gst_gl_filterblur_hcallback (gint width, gint height, guint texture, gpointer st
   gst_gl_shader_set_uniform_1i (filterblur->shader0, "tex", 1);
 
   gst_gl_shader_set_uniform_1fv (filterblur->shader0, "kernel", 9, gauss_kernel);
-  gst_gl_shader_set_uniform_1f (filterblur->shader0, "norm_const", 0.977016);
-  gst_gl_shader_set_uniform_1f (filterblur->shader0, "norm_offset", 0.0);
+  gst_gl_shader_set_uniform_1f (filterblur->shader0, "norm_const", 0.977016f);
+  gst_gl_shader_set_uniform_1f (filterblur->shader0, "norm_offset", 0.0f);
 
   gst_gl_filterblur_draw_texture (filterblur, texture);
 }
@@ -322,9 +322,9 @@ gst_gl_filterblur_vcallback (gint width, gint height, guint texture, gpointer st
   /* hard coded kernel, it could be easily generated at runtime with a
    * property to change standard deviation */
   gfloat gauss_kernel[9] = {
-    0.026995, 0.064759, 0.120985,
-    0.176033, 0.199471, 0.176033,
-    0.120985, 0.064759, 0.026995 };
+    0.026995f, 0.064759f, 0.120985f,
+    0.176033f, 0.199471f, 0.176033f,
+    0.120985f, 0.064759f, 0.026995f };
 
   glMatrixMode (GL_PROJECTION);
   glLoadIdentity ();
@@ -339,8 +339,8 @@ gst_gl_filterblur_vcallback (gint width, gint height, guint texture, gpointer st
   gst_gl_shader_set_uniform_1i (filterblur->shader1, "tex", 1);
 
   gst_gl_shader_set_uniform_1fv (filterblur->shader1, "kernel", 9, gauss_kernel);
-  gst_gl_shader_set_uniform_1f (filterblur->shader1, "norm_const", 0.977016);
-  gst_gl_shader_set_uniform_1f (filterblur->shader1, "norm_offset", 0.0);
+  gst_gl_shader_set_uniform_1f (filterblur->shader1, "norm_const", 0.977016f);
+  gst_gl_shader_set_uniform_1f (filterblur->shader1, "norm_offset", 0.0f);
 
   gst_gl_filterblur_draw_texture (filterblur, texture);
 }
